+++ src/controller.h
@@ -93,7 +93,7 @@ private:
     void refreshSentZTrans();
     void refreshReceivedZTrans(QList<QString> zaddresses);
 
-    bool processUnspent     (const json& reply, QMap<QString, double>* newBalances, QList<UnspentOutput>* newUtxos);
+    bool processUnspent     (const json& reply, QMap<QString, qint64>* newBalances, QList<UnspentOutput>* newUtxos);
     void updateUI           (bool anyUnconfirmed);
 
     void getInfoThenRefresh(bool force);
